+++ src/commands/index.ts
@@ -3,3 +3,4 @@ export { createSummaryPerRunCommand } from "./summaryPerRun";
 export { createLineCommand, createBarCommand } from "./lineBar";
 export { createBoxPlotCommand } from "./boxplot";
 export { createTableCommand } from "./table";
+export { createEntityBreakdownCommand } from "./entityBreakdown";